+++ graphannis/src/annis/db/corpusstorage/tests.rs
@@ -2,6 +2,7 @@ extern crate log;
 extern crate tempfile;
 
 use std::path::PathBuf;
+use std::vec;
 
 use crate::annis::db::corpusstorage::get_read_or_error;
 use crate::annis::db::{aql::model::AnnotationComponentType, example_generator};
@@ -239,6 +240,112 @@ fn subgraph_with_segmentation() {
     assert_eq!(5, seg2_out.unwrap().len());
 
     assert_eq!(None, graph.get_node_id_from_name("root/doc1#seg3").unwrap());
+}
+
+/// Test that context generation works with a corpus that has segmentations
+/// and gaps in the segments.
+#[test]
+fn subgraph_with_segmentation_and_gap() {
+    let tmp = tempfile::tempdir().unwrap();
+    let cargo_dir = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
+
+    let cs = CorpusStorage::with_auto_cache_size(tmp.path(), true).unwrap();
+    let corpus_name = cs
+        .import_from_fs(
+            &cargo_dir.join("tests/SegmentationWithGaps.graphml"),
+            ImportFormat::GraphML,
+            None,
+            false,
+            true,
+            |_| {},
+        )
+        .unwrap();
+
+    // Use the norm="Gaps" node as match which is an existing segmentation
+    let m = vec!["SegmentationWithGaps/doc01#norm12".to_string()];
+
+    // Get the context using tokens
+    let g = cs.subgraph(&corpus_name, m.clone(), 1, 2, None).unwrap();
+    // Check that all token and the page are included, including the token
+    // that is not covered by a segmentation node.
+    assert!(g
+        .get_node_id_from_name("SegmentationWithGaps/doc01#tok_11")
+        .unwrap()
+        .is_some());
+    assert!(g
+        .get_node_id_from_name("SegmentationWithGaps/doc01#tok_12")
+        .unwrap()
+        .is_some());
+    assert!(g
+        .get_node_id_from_name("SegmentationWithGaps/doc01#tok_13")
+        .unwrap()
+        .is_some());
+    assert!(g
+        .get_node_id_from_name("SegmentationWithGaps/doc01#tok_14")
+        .unwrap()
+        .is_some());
+    assert!(g
+        .get_node_id_from_name("SegmentationWithGaps/doc01#page2")
+        .unwrap()
+        .is_some());
+
+    // Get the context for the norm node using the norm segmentation
+    let g = cs
+        .subgraph(&corpus_name, m, 1, 1, Some("norm".to_string()))
+        .unwrap();
+    // Check that all token and the page are included
+    assert!(g
+        .get_node_id_from_name("SegmentationWithGaps/doc01#tok_11")
+        .unwrap()
+        .is_some());
+    assert!(g
+        .get_node_id_from_name("SegmentationWithGaps/doc01#tok_12")
+        .unwrap()
+        .is_some());
+    assert!(g
+        .get_node_id_from_name("SegmentationWithGaps/doc01#tok_13")
+        .unwrap()
+        .is_some());
+    assert!(g
+        .get_node_id_from_name("SegmentationWithGaps/doc01#tok_14")
+        .unwrap()
+        .is_some());
+    assert!(g
+        .get_node_id_from_name("SegmentationWithGaps/doc01#page2")
+        .unwrap()
+        .is_some());
+
+    // Get the context for the token using the norm segmentation
+    let g = cs
+        .subgraph(
+            &corpus_name,
+            vec!["SegmentationWithGaps/doc01#tok_12".to_string()],
+            1,
+            1,
+            Some("norm".to_string()),
+        )
+        .unwrap();
+    // Check that all token and the page are included
+    assert!(g
+        .get_node_id_from_name("SegmentationWithGaps/doc01#tok_11")
+        .unwrap()
+        .is_some());
+    assert!(g
+        .get_node_id_from_name("SegmentationWithGaps/doc01#tok_12")
+        .unwrap()
+        .is_some());
+    assert!(g
+        .get_node_id_from_name("SegmentationWithGaps/doc01#tok_13")
+        .unwrap()
+        .is_some());
+    assert!(g
+        .get_node_id_from_name("SegmentationWithGaps/doc01#tok_14")
+        .unwrap()
+        .is_some());
+    assert!(g
+        .get_node_id_from_name("SegmentationWithGaps/doc01#page2")
+        .unwrap()
+        .is_some());
 }
 
 fn compare_annos<T>(